Got DCEL text output working.
Got the DCEL text output to look correct, did some very simple testing but it looks to be working. Started on the graphics portion of the project, got the window set up and parsed in the vertices and drew them to the window. Next step is parsing in the hedges and printing them to the window.